Rename Page and Entity
