+++ src/EthernaIndex.Persistence/ModelMaps/Index/UnsuitableReportBaseMap.cs
@@ -1,12 +1,11 @@
 ﻿using Etherna.EthernaIndex.Domain.Models;
-using Etherna.EthernaIndex.Persistence.ModelMaps.Index;
 using Etherna.MongODM.Core;
 using Etherna.MongODM.Core.Extensions;
 using Etherna.MongODM.Core.Serialization;
 
 namespace Etherna.EthernaIndex.Persistence.ModelMaps.Index
 {
-    class ReportBaseMap : IModelMapsCollector
+    class UnsuitableReportBaseMap : IModelMapsCollector
     {
         public void Register(IDbContext dbContext)
         {
