+++ src/EthernaIndex/Areas/Admin/Pages/Index.cshtml
@@ -25,8 +25,8 @@
 
         <ul>
             <li><a class="btn btn-link" asp-area="@Etherna.EthernaIndex.Configs.CommonConsts.AdminArea" asp-page="/VideoManifests/Index">Video Manifests</a></li>
-            <li><a class="btn btn-link" asp-area="@Etherna.EthernaIndex.Configs.CommonConsts.AdminArea" asp-page="/VideoReports/Index">Video Reports</a></li>
-            <li><a class="btn btn-link" asp-area="@Etherna.EthernaIndex.Configs.CommonConsts.AdminArea" asp-page="/VideoReviews/Index">Video Reviews</a></li>
+            <li><a class="btn btn-link" asp-area="@Etherna.EthernaIndex.Configs.CommonConsts.AdminArea" asp-page="/VideoUnsuitableReports/Index">Video Unsuitable Reports</a></li>
+            <li><a class="btn btn-link" asp-area="@Etherna.EthernaIndex.Configs.CommonConsts.AdminArea" asp-page="/VideoUnsuitableReviews/Index">Video Unsuitable Reviews</a></li>
         </ul>
 
     </div>
