+++ src/EthernaIndex/Areas/Admin/Pages/VideoUnsuitableReviews/Index.cshtml
@@ -14,7 +14,7 @@
 *@
 
 @page
-@model Etherna.EthernaIndex.Areas.Admin.Pages.Reviews.IndexModel
+@model Etherna.EthernaIndex.Areas.Admin.Pages.VideoUnsuitableReviews.IndexModel
 @{
     ViewData["Title"] = "Video Reviews";
