+++ src/EthernaIndex/Areas/Admin/Pages/VideoUnsuitableReviews/Index.cshtml.cs
@@ -10,17 +10,14 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Etherna.EthernaIndex.Areas.Admin.Pages.VideoReports
+namespace Etherna.EthernaIndex.Areas.Admin.Pages.VideoUnsuitableReviews
 {
-    public class ReportModel : PageModel
+    public class IndexModel : PageModel
     {
 
         // Models.
         public class InputModel
         {
-            [Display(Name = "Include Reviewed")]
-            public bool IncludeReportReviewed { get; set; } = default!;
-
             [Display(Name = "Manifest Hash")]
             public string? ManifestHash { get; set; }
 
@@ -28,30 +25,27 @@ namespace Etherna.EthernaIndex.Areas.Admin.Pages.VideoReports
             public string? VideoId { get; set; }
         }
 
-        public class VideoReportDto
+        public class VideoReviewDto
         {
-            public VideoReportDto(
-                string manifestHash,
+            public VideoReviewDto(
+                string lastValidManifestHash,
                 string title,
-                string videoId,
-                int totalReports)
+                string videoId)
             {
-                if (manifestHash is null)
-                    throw new ArgumentNullException(nameof(manifestHash));
+                if (lastValidManifestHash is null)
+                    throw new ArgumentNullException(nameof(lastValidManifestHash));
                 if (title is null)
                     throw new ArgumentNullException(nameof(title));
                 if (videoId is null)
                     throw new ArgumentNullException(nameof(videoId));
 
-                ManifestHash = manifestHash;
+                LastValidManifestHash = lastValidManifestHash;
                 Title = title;
-                TotalReports = totalReports;
                 VideoId = videoId;
             }
 
-            public string ManifestHash { get; }
+            public string LastValidManifestHash { get; }
             public string Title { get; }
-            public int TotalReports { get; }
             public string VideoId { get; }
         }
 
@@ -62,7 +56,7 @@ namespace Etherna.EthernaIndex.Areas.Admin.Pages.VideoReports
         private readonly IIndexDbContext indexDbContext;
 
         // Constructor.
-        public ReportModel(
+        public IndexModel(
             IIndexDbContext indexDbContext)
         {
             this.indexDbContext = indexDbContext;
@@ -74,17 +68,15 @@ namespace Etherna.EthernaIndex.Areas.Admin.Pages.VideoReports
 
         public int CurrentPage { get; private set; }
         public int MaxPage { get; private set; }
-        public IEnumerable<VideoReportDto> VideoReports { get; private set; } = default!;
+        public IEnumerable<VideoReviewDto> VideoReports { get; private set; } = default!;
 
         // Methods.
         public async Task OnGetAsync(
-            bool includeReportReviewed,
-            string manifestHash,
-            string videoId,
+            string? manifestHash,
+            string? videoId,
             int? p)
         {
             await InitializeAsync(
-                includeReportReviewed,
                 manifestHash,
                 videoId,
                 p);
@@ -92,16 +84,15 @@ namespace Etherna.EthernaIndex.Areas.Admin.Pages.VideoReports
 
         // Helpers.
         private async Task InitializeAsync(
-            bool includeReportReviewed,
             string? manifestHash,
             string? videoId,
             int? p)
         {
             CurrentPage = p ?? 0;
 
-            var paginatedVideos = await indexDbContext.VideoUnsuitableReports.QueryPaginatedElementsAsync(
-                vm => VideoReportsWhere(vm, includeReportReviewed, manifestHash, videoId)
-                        .GroupBy(i => i.VideoManifest.Id)
+            var paginatedVideoReviews = await indexDbContext.VideoUnsuitableReviews.QueryPaginatedElementsAsync(
+                vm => VideoReviewsWhere(vm, manifestHash, videoId)
+                        .GroupBy(i => i.VideoId)
                         .Select(group => new
                         {
                             Id = group.Key,
@@ -111,44 +102,40 @@ namespace Etherna.EthernaIndex.Areas.Admin.Pages.VideoReports
                 CurrentPage,
                 PageSize);
 
-            MaxPage = paginatedVideos.MaxPage;
+            MaxPage = paginatedVideoReviews.MaxPage;
 
-            var videoManifestIds = paginatedVideos.Elements.Select(e => e.Id);
+            var videoReviewsIds = paginatedVideoReviews.Elements.Select(e => e.Id);
 
-            // Get manifest info.
-            var videoManifests = await indexDbContext.VideoManifests.QueryElementsAsync(elements =>
-               elements.Where(u => videoManifestIds.Contains(u.Id))
-                       .OrderBy(i => i.Id)
-                       .ToListAsync());
+            // Get video and manifest info.
+            var videos = await indexDbContext.Videos.QueryElementsAsync(elements =>
+                elements.Where(u => videoReviewsIds.Contains(u.Id))
+                        .OrderBy(i => i.Id)
+                        .ToListAsync());
 
-            VideoReports = videoManifests.Select(vm => new VideoReportDto(
-                vm.ManifestHash.Hash,
-                vm.Title ?? "",
-                vm.Video.Id,
-                paginatedVideos.Elements.First(pv => pv.Id == vm.Id).Count));
+            VideoReports = videos.Select(v =>
+            {
+                var manifest = v.GetLastValidManifest();
+                return new VideoReviewDto(manifest?.ManifestHash?.Hash ?? "", manifest?.Title ?? "", v.Id);
+            });
         }
 
         public async Task OnPost()
         {
             await InitializeAsync(
-                Input?.IncludeReportReviewed ?? false,
                 Input?.ManifestHash,
                 Input?.VideoId,
                 null);
         }
 
-        private IMongoQueryable<VideoUnsuitableReport> VideoReportsWhere(
-            IMongoQueryable<VideoUnsuitableReport> querable,
-            bool includeReportReviewed,
+        private IMongoQueryable<VideoUnsuitableReview> VideoReviewsWhere(
+            IMongoQueryable<VideoUnsuitableReview> querable,
             string? manifestHash,
             string? videoId)
         {
             if (!string.IsNullOrWhiteSpace(manifestHash))
-                querable = querable.Where(vr => vr.VideoManifest.ManifestHash.Hash == manifestHash);
+                querable = querable.Where(vr => vr.ManifestHash == manifestHash);
             if (!string.IsNullOrWhiteSpace(videoId))
-                querable = querable.Where(vr => vr.VideoManifest.Video.Id == videoId);
-            if (includeReportReviewed)
-                return querable.Where(vr => vr.VideoManifest.ReviewApproved == null);
+                querable = querable.Where(vr => vr.VideoId == videoId);
 
             return querable;
         }
